requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
altair>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
